refactor: rename defuse.ts -> authIdentity.ts
