+++ src/features/gift/hooks/useBalanceUpdaterSyncWithHistory.ts
@@ -1,7 +1,7 @@
 import { useEffect, useRef } from "react"
 import type { ActorRefFrom } from "xstate"
 import type { SignerCredentials } from "../../../core/formatters"
-import { authHandleToIntentsUserId } from "../../../utils/defuse"
+import { authHandleToIntentsUserId } from "../../../utils/authIdentity"
 import type { giftMakerRootMachine } from "../actors/giftMakerRootMachine"
 import {
   type GiftMakerHistory,
