+++ src/features/gift/utils/determineGiftToken.ts
@@ -2,7 +2,7 @@ import { Err, Ok, type Result } from "@thames/monads"
 import { nearClient } from "../../../constants/nearClient"
 import { getDepositedBalances } from "../../../services/defuseBalanceService"
 import type { BaseTokenInfo, UnifiedTokenInfo } from "../../../types/base"
-import { authHandleToIntentsUserId } from "../../../utils/defuse"
+import { authHandleToIntentsUserId } from "../../../utils/authIdentity"
 import { isBaseToken, isUnifiedToken } from "../../../utils/token"
 import { getUnderlyingBaseTokenInfos } from "../../../utils/tokenUtils"
 import type { EscrowCredentials } from "./generateEscrowCredentials"
