+++ src/features/gift/utils/signGiftTakerMessage.ts
@@ -1,6 +1,6 @@
 import { base64 } from "@scure/base"
 import { KeyPair } from "near-api-js"
-import { authHandleToIntentsUserId } from "src/utils/defuse"
+import { authHandleToIntentsUserId } from "src/utils/authIdentity"
 import type { IntentsUserId, SignerCredentials } from "../../../core/formatters"
 import { formatUserIdentity } from "../../../core/formatters"
 import type { NEP413SignatureData } from "../../../types/swap"
