+++ src/features/machines/swapUIMachine.ts
@@ -18,7 +18,7 @@ import type {
 } from "../../types/base"
 import type { SwappableToken } from "../../types/swap"
 import { assert } from "../../utils/assert"
-import { authHandleToIntentsUserId } from "../../utils/defuse"
+import { authHandleToIntentsUserId } from "../../utils/authIdentity"
 import { parseUnits } from "../../utils/parse"
 import {
   getAnyBaseTokenInfo,
