+++ src/features/otcDesk/stores/otcMakerTrades.ts
@@ -3,7 +3,7 @@ import { createJSONStorage, persist } from "zustand/middleware"
 import type { SignerCredentials } from "../../../core/formatters"
 import type { MultiPayload } from "../../../types/defuse-contracts-types"
 import type { IntentsUserId } from "../../../types/intentsUserId"
-import { authHandleToIntentsUserId } from "../../../utils/defuse"
+import { authHandleToIntentsUserId } from "../../../utils/authIdentity"
 
 type OtcMakerTrade = {
   tradeId: string
