+++ src/services/depositService.ts
@@ -36,7 +36,7 @@ import type { SendTransactionEVMParams, Transaction } from "../types/deposit"
 import type { IntentsUserId } from "../types/intentsUserId"
 import { BlockchainEnum } from "../types/interfaces"
 import { assert } from "../utils/assert"
-import { authHandleToIntentsUserId } from "../utils/defuse"
+import { authHandleToIntentsUserId } from "../utils/authIdentity"
 import { getEVMChainId } from "../utils/evmChainId"
 import { isNativeToken } from "../utils/token"
 import { getDepositAddress, getSupportedTokens } from "./poaBridgeClient"
